Time: 3 ms (66.80%), Space: 6 MB (26.83%) - LeetHub

+++ 69-sqrtx/69-sqrtx.cpp
@@ -5,7 +5,8 @@ public:
         int hi = INT_MAX;
         int ans = 0;
         while(lo<=hi){
-            long long mid = hi + (lo - hi)/2;
+            
+            double mid = hi + (lo - hi)/2;
             if(mid*mid <= x) {
                 ans = mid;
                 lo = mid + 1;
